backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.8,<3.0
python-dotenv>=1.0,<2.0
PyMuPDF>=1.24,<2.0
chromadb>=1.0,<2.0
langchain>=1.0,<2.0
langchain-core>=1.0,<2.0
langchain-classic>=1.0,<2.0
langchain-community>=0.4,<1.0
langchain-chroma>=1.0,<2.0
langchain-huggingface>=1.0,<2.0
langchain-text-splitters>=1.0,<2.0
langchain-deepseek>=1.0,<2.0
sentence-transformers>=3.0,<5.0
requests>=2.31,<3.0